legacy_version/web/requirements.txt
flask
cryptography
bleach
djlint>=1.32.1
gunicorn
google-cloud-dialogflow
sqlalchemy
flask_sqlalchemy
psycopg2-binary
python-dotenv
user_agents